feat: add translations for app metadata (#33)
This PR implements the [app metadata translation
approach](https://docs.expo.dev/guides/localization/#translating-app-metadata)
recommended by Expo.

I tested with my device language set to French, and verified that the
custom strings are used:
![Simulator Screenshot - iPhone 15 Pro Max - 2025-03-17 at 17 06
13](https://github.com/user-attachments/assets/975331a9-6dec-46fd-a02e-5940830887d2)

One thing to note with this implementation is at compile / build time,
expo takes the provided locales files and copies them to the relevant
languages project's InfoPlist.strings file. Any custom translations
added to the locales files will also be copied here unnecessarily - this
is only a problem if there are so many custom translations that it
bloats the app size / affects app startup. For now I think it is okay,
but wanted to make a note.

Related to ENG-196